updated to use new constructor

+++ Test Programs/PressureSensorCalibration/lib/pressureSensor/pressureSensor.h
@@ -19,7 +19,7 @@ class PressureSensor {
     float pressure;
 
    public:
-    PressureSensor(float sampleSize_,float pressureVariation_, float depthVariation_, float tempVariation_, int fluidDensity_)
+    PressureSensor(float sampleSize_,float pressureVariation_, float depthVariation_, float tempVariation_, float fluidDensity_)
     :sampleSize(sampleSize_),
     pressureVariation(pressureVariation_),
     depthVariation(depthVariation_),
